+++ mapie/tests/test_calibration.py
@@ -157,11 +157,28 @@ def test_other_methods() -> None:
         mapie_cal.fit(X, y)
 
 
-def test_prefit() -> None:
+def test_prefit_cv_argument() -> None:
     """Test that prefit method works"""
     est = RandomForestClassifier().fit(X, y)
     mapie_cal = MapieCalibrator(estimator=est, cv="prefit")
     mapie_cal.fit(X, y)
+
+
+def test_split_cv_argument() -> None:
+    """Test that split method works"""
+    mapie_cal = MapieCalibrator(cv="split")
+    mapie_cal.fit(X, y)
+
+
+@pytest.mark.parametrize("cv", ["noprefit", "nosplit"])
+def test_invalid_cv_argument(cv: str) -> None:
+    """Test that other cv method does not work"""
+    with pytest.raises(
+        ValueError,
+        match=r".*Invalid cv argument*",
+    ):
+        mapie_cal = MapieCalibrator(cv=cv)
+        mapie_cal.fit(X, y)
 
 
 def test_prefit_split_same_results() -> None: